refactor: :recycle: create card component collection
