+++ lib/view/home/collection/my_collections.dart
@@ -1,8 +1,9 @@
+import 'package:bookclub/common/collection_card.dart';
 import 'package:bookclub/common/loader.dart';
+import 'package:bookclub/common/style_manager.dart';
 import 'package:bookclub/model/collection.dart';
 import 'package:bookclub/repository/collection.dart';
-import 'package:bookclub/routes/app_routes.dart';
-import 'package:bookclub/view/home/collection/collection_books.dart';
+import 'package:bookclub/view/home/discontinued_pages/edit_book.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_slidable/flutter_slidable.dart';
 
@@ -66,6 +67,13 @@ class _MyCollectionPageState extends State<MyCollectionPage> {
           ),
         ),
       ),
+      floatingActionButton: FloatingActionButton(
+        onPressed: () {
+          _goToEditBook();
+        },
+        backgroundColor: StyleManager.instance.primary,
+        child: Icon(Icons.add),
+      ),
       body: Container(
           color: Colors.grey.shade900,
           child: _isLoading
@@ -103,13 +111,19 @@ class _MyCollectionPageState extends State<MyCollectionPage> {
                               ),
                             ],
                           ),
-                          child: collectionComponent(
-                              collection: _foundedMyCollections[index],
-                              context: context),
-                        );
+                          child: CollectionCard(
+                            collection: _foundedMyCollections[index]),
+                      );
                       })
                   : const Center(child: Text("Nenhuma coleção sua foi encontrada"))),
     );
+  }
+
+  _goToEditBook() {
+    Navigator.push(
+            context,
+            MaterialPageRoute(builder: (context) => EditBookPage()),
+          );
   }
 
   _getMyCollections() async {
@@ -129,50 +143,6 @@ class _MyCollectionPageState extends State<MyCollectionPage> {
     }
     setState(() => _isLoading = false);
   }
-}
-
-collectionComponent({required Collection collection, context}) {
-  return GestureDetector(
-    onTap: () {
-      Navigator.pushNamed(context, AppRoutes.COLLECTION_BOOKS,
-          arguments: collection);
-    },
-    child: Padding(
-      padding: const EdgeInsets.all(10),
-      child: Container(
-        padding: const EdgeInsets.all(20),
-        //color: Colors.amber,
-        decoration: BoxDecoration(
-            border: Border.all(
-              color: Colors.white,
-            ),
-            borderRadius: const BorderRadius.all(Radius.circular(20))),
-        child: Column(
-          children: [
-            Row(
-              children: [
-                //Text(collection.image),
-                //const Icon(Icons.book_outlined, color: Colors.white, size: 50),
-                //SizedBox(width: 10),
-                Text(
-                  collection.title ?? '',
-                  style: const TextStyle(
-                      color: Colors.white,
-                      fontSize: 20,
-                      fontWeight: FontWeight.bold),
-                ),
-              ],
-            ),
-            const SizedBox(height: 10),
-            Text(
-              collection.description ?? '',
-              style: const TextStyle(color: Colors.white),
-            ),
-          ],
-        ),
-      ),
-    ),
-  );
 }
 
 void doNothing(BuildContext context) {}
